Add: rough draft of holochain package mgmt presentation
Holochain will need a sane method to manage packages and dependancies
deterministically because of the nature of the project's goals. The DNA of an
app is determined by the components of a fuction defined in the zome. A package
management tool (and language) which is deterministic will be key to realizing
this efficiently and effectively.